requirements.txt
asgiref==3.5.2
Babel==2.10.3
certifi==2022.6.15
charset-normalizer==2.1.1
Django==3.2.8
django-debug-toolbar==3.6.0
django-extensions==3.2.0
django-multi-form-view==2.0.1
django-phonenumber-field==6.3.0
django-widget-tweaks==1.4.12
graphviz==0.20.1
idna==3.3
MarkupSafe==2.1.1
numpy==1.22.3
opencv-python==4.5.5.64
phonenumberslite==8.12.54
Pillow==8.4.0
psycopg2==2.9.3
pydotplus==2.0.2
pyparsing==3.0.9
pytz==2022.1
requests==2.28.1
sqlparse==0.4.2
urllib3==1.26.12
Werkzeug==2.2.2
